Fix pr suggestions
still a few more to resolve
